+++ assets/js/geckopress-base.js
@@ -1,9 +1,7 @@
 function hover_dropdown_fix()
 {
-    console.log('dropdown fixing');
     if(jQuery(window).width() <= 768)
     {
-        console.log('mobile');
         jQuery('.navbar-nav.hover-dropdown.mobile-multilevel .menu-item-has-children > a')
             .addClass('dropdown-toggle')
             .attr('data-toggle','dropdown');