+++ so-widgets/hello-world-widget/tpl/hello-world-template.php
@@ -1,0 +1,3 @@
+<div class="hello-world">
+    <?php echo wp_kses_post($instance['text']) ?>
+</div>