+++ changelog.md
@@ -2,7 +2,6 @@
 
 ## 0.2
 
-- Lot of bugfixes and improvements
 - Bootstrap .container css class renamed to bs-container because of a collision with SiteOrigin Page Builder
 - Better code quality
 - Default YAML files changed (non associative arrays if not needed)
@@ -10,7 +9,9 @@
 - Added hover dropdown navigation menus optionally
 - Responsive video embed added
 - GeckoPress base js added
-
+- Contact Form 7 widget for SiteOrigin Page Builder added.
+- Lot of other bugfixes and improvements
+- 
 ## 0.1.1
 
 - Lot of bugfixes and improvements